requirements.txt
drawsvg[all]
Flask==2.2.2
Pillow==10.0.0
svglib
rlPyCairo